requirements-2.9.txt
#
# This file is autogenerated by pip-compile with python 3.10
# 
# It has been pinned to test in environments where
# there are juju controllers versioned either with 
# 2.9.x or 3.0.x
#
async-generator==1.10
    # via
    #   -r requirements.in
    #   asyncio-extras
asyncio-extras==1.3.2
    # via -r requirements.in
attrs==23.1.0
    # via pytest
awscli==1.27.99
    # via -r requirements.in
backoff==2.2.1
    # via -r requirements.in
bcrypt==4.0.1
    # via paramiko
beautifulsoup4==4.12.2
    # via bs4
boto3==1.26.99
    # via -r requirements.in
botocore==1.29.99
    # via
    #   awscli
    #   boto3
    #   s3transfer
bs4==0.0.1
    # via -r requirements.in
cachetools==5.3.3
    # via google-auth
certifi==2023.7.22
    # via
    #   kubernetes
    #   requests
cffi==1.15.1
    # via
    #   bcrypt
    #   cryptography
    #   pynacl
charset-normalizer==3.1.0
    # via requests
click==8.1.6
    # via
    #   mkdocs
colorama==0.4.4
    # via
    #   awscli
    #   mkdocs-material
cryptography==42.0.5
    # via
    #   paramiko
decorator==5.1.1
    # via retry
dill==0.3.6
    # via -r requirements.in
distro==1.8.0
    # via lazr-restfulclient
docutils==0.16
    # via awscli
drypy==1.0.1
    # via -r requirements.in
exceptiongroup==1.1.2
    # via pytest
flake8==6.0.0
    # via -r requirements.in
flaky==3.7.0
    # via -r requirements.in
ghp-import==2.1.0
    # via mkdocs
google-auth==2.23.4
    # via kubernetes
httplib2==0.22.0
    # via
    #   launchpadlib
    #   lazr-restfulclient
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
invoke==2.1.3
    # via -r requirements.in
iso8601==2.1.0
    # via surl
jinja2==3.1.3
    # via
    #   -r requirements.in
    #   mkdocs
    #   mkdocs-material
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
juju==3.0.4
    # via -r requirements.in
jujubundlelib==0.5.7
    # via theblues
kubernetes==26.1.0
    # via juju
kv==0.4.1
    # via -r requirements.in
launchpadlib==1.11.0
    # via -r requirements.in
lazr-restfulclient==0.14.5
    # via
    #   -r requirements.in
    #   launchpadlib
lazr-uri==1.0.6
    # via
    #   launchpadlib
    #   wadllib
loguru==0.7.0
    # via -r requirements.in
macaroonbakery==1.3.1
    # via
    #   juju
    #   theblues
markdown==3.3.7
    # via
    #   mkdocs
    #   mkdocs-material
    #   pymdown-extensions
markupsafe==2.1.3
    # via jinja2
mccabe==0.7.0
    # via flake8
mergedeep==1.3.4
    # via mkdocs
mkdocs==1.4.3
    # via
    #   -r requirements.in
    #   mkdocs-material
mkdocs-material==9.1.18
    # via -r requirements.in
mkdocs-material-extensions==1.1.1
    # via mkdocs-material
mypy-extensions==1.0.0
    # via
    #   typing-inspect
oauthlib==3.2.2
    # via
    #   lazr-restfulclient
    #   requests-oauthlib
packaging==23.2
    # via
    #   mkdocs
    #   pytest
paramiko==2.12.0
    # via juju
pbr==5.11.1
    # via
    #   stevedore
pluggy==1.2.0
    # via pytest
prettytable==3.7.0
    # via -r requirements.in
protobuf==3.20.2
    # via macaroonbakery
py==1.11.0
    # via
    #   pytest-html
    #   retry
pyasn1==0.5.0
    # via
    #   juju
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pycodestyle==2.10.0
    # via flake8
pycparser==2.21
    # via cffi
pyflakes==3.0.1
    # via flake8
pygments==2.16.1
    # via mkdocs-material
pymacaroons==0.13.0
    # via
    #   -r requirements.in
    #   macaroonbakery
    #   surl
pymdown-extensions==10.3.1
    # via mkdocs-material
pynacl==1.5.0
    # via
    #   macaroonbakery
    #   paramiko
    #   pymacaroons
pyparsing==3.0.9
    # via httplib2
pyrfc3339==1.1
    # via
    #   juju
    #   macaroonbakery
pytest==7.4.0
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-html
    #   pytest-json-report
    #   pytest-metadata
    #   pytest-mock
pytest-asyncio==0.21.0
    # via -r requirements.in
pytest-html==3.2.0
    # via -r requirements.in
pytest-json-report==1.5.0
    # via -r requirements.in
pytest-metadata==3.0.0
    # via
    #   pytest-html
    #   pytest-json-report
pytest-mock==3.11.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   ghp-import
    #   kubernetes
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3
    # via pyrfc3339
pyyaml==5.4.1
    # via
    #   -r requirements.in
    #   awscli
    #   juju
    #   jujubundlelib
    #   kubernetes
    #   mkdocs
    #   pyyaml-env-tag
pyyaml-env-tag==0.1
    # via mkdocs
regex==2023.8.8
    # via mkdocs-material
requests==2.31.0
    # via
    #   -r requirements.in
    #   kubernetes
    #   macaroonbakery
    #   mkdocs-material
    #   requests-oauthlib
    #   surl
    #   theblues
requests-oauthlib==1.3.1
    # via kubernetes
retry==0.9.2
    # via -r requirements.in
rsa==4.7.2
    # via
    #   awscli
    #   google-auth
s3transfer==0.6.1
    # via
    #   awscli
    #   boto3
semver==2.13.0
    # via -r requirements.in
sh==2.0.4
    # via -r requirements.in
six==1.16.0
    # via
    #   bcrypt
    #   fasteners
    #   google-auth
    #   kubernetes
    #   launchpadlib
    #   lazr-restfulclient
    #   macaroonbakery
    #   paramiko
    #   pymacaroons
    #   pynacl
    #   python-dateutil
soupsieve==2.4.1
    # via beautifulsoup4
surl==0.7.2
    # via -r requirements.in
tabulate==0.9.0
    # via surl
theblues==0.5.2
    # via juju
tomli==2.0.1
    # via
    #   pytest
toposort==1.10
    # via juju
typing-extensions==4.7.1
    # via typing-inspect
typing-inspect==0.9.0
    # via juju
urllib3==1.26.15
    # via
    #   botocore
    #   kubernetes
    #   requests
wadllib==1.3.6
    # via lazr-restfulclient
watchdog==3.0.0
    # via mkdocs
wcwidth==0.2.6
    # via prettytable
websocket-client==1.5.1
    # via kubernetes
websockets==7.0 ; python_version < "3.9"
websockets==8.1 ; python_version == "3.9"
websockets==11.0.3 ; python_version >= "3.10"
    # via
    #   -r requirements.in
    #   juju
wheel==0.40.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
